Refactored authentication errors into authentication
