+++ blackrose-website-backend/src/authentication.rs
@@ -1,4 +1,5 @@
 pub mod auth;
+pub mod errors;
 pub mod login;
-pub mod validation;
 pub mod registration;
+pub mod validation;
