+++ blackrose-website-backend/src/authentication/login.rs
@@ -1,10 +1,10 @@
 use crate::{
+    authentication::errors::LoginError,
     authentication::{
         auth::{expire_in_five_hours, Claims},
         validation::UserCredentials,
     },
     database::models::User,
-    errors::LoginError,
     state::AppState,
     KEYS,
 };
